Move RoutingManager out of VehicleManager
