+++ src/OpenLoco/CMakeLists.txt
@@ -216,6 +216,7 @@ set(OLOCO_SOURCES
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Vehicles/OrderSkip.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Vehicles/Orders.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Vehicles/RenameVehicle.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Vehicles/RoutingManager.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Vehicles/Routing.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Vehicles/Vehicle.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Vehicles/Vehicle1.cpp"
@@ -470,6 +471,7 @@ set(OLOCO_HEADERS
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/WindowType.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Vehicles/OrderManager.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Vehicles/Orders.h"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Vehicles/RoutingManager.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Vehicles/Routing.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Vehicles/Vehicle.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Vehicles/VehicleManager.h"
