+++ src/OpenLoco/src/Vehicles/CreateVehicle.cpp
@@ -20,6 +20,7 @@
 #include "Vehicle.h"
 #include "VehicleManager.h"
 #include "Vehicles/OrderManager.h"
+#include "Vehicles/RoutingManager.h"
 #include "World/CompanyManager.h"
 #include "World/Station.h"
 #include <numeric>
