+++ src/OpenLoco/src/Vehicles/Vehicle1.cpp
@@ -1,7 +1,7 @@
 #include "Objects/BridgeObject.h"
 #include "Objects/RoadObject.h"
+#include "RoutingManager.h"
 #include "Vehicle.h"
-#include "VehicleManager.h"
 #include <OpenLoco/Interop/Interop.hpp>
 
 using namespace OpenLoco::Interop;
