+++ src/OpenLoco/src/Vehicles/VehicleTail.cpp
@@ -4,8 +4,8 @@
 #include "Map/TileManager.h"
 #include "Map/Track/Track.h"
 #include "Map/Track/TrackData.h"
+#include "RoutingManager.h"
 #include "Vehicle.h"
-#include "VehicleManager.h"
 #include "ViewportManager.h"
 
 using namespace OpenLoco::Interop;
